+++ src/main/java/com/example/travelday/global/config/WebConfig.java
@@ -1,8 +1,6 @@
-package com.example.travelday.global.config.security;
+package com.example.travelday.global.config;
 
-import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
